Stub out lastFlipResultDescription and associated label
Add -lastFlipResultDescription to SCSCardMatchingGame, which when
called will return a description of whatever the most recent action in
the game is. Initially, this has been implemented to return nil if
nothing has happened, or return "Game has started" if the game has
started.

The game has started once the first card is flipped. Implemented using
the private @started property:

    @property (nonatomic, getter=isStarted) BOOL started;