Implement real descriptions of the last flip results
Implement -setLastFlipResult: and -setLastFlipResult:forCards:score:
setters which record all the relevant info for a given last result.
Conceptually decided to treat this as still being a property, but with
extra information if necessary which is stored in new, hidden instance
variables.

Building the description in a switch statement based on all the last
result info that is stored in the matching algorithm.